requirements.txt
PyOpenGL
PyOpenGL_accelerate
numpy
pygame
fastapi
pydantic
typing
uvicorn
requests

